Custom tile type and editor
Custom tile that has a footstep color value which the player particle system uses to determine the color of the dust. Create a new spritesheet for the custom tile and split it. Open the new custom editor and drag the spritesheet and create a new tile palette from the create menu and drag that as well. Set the correct tile size and click process. Choose a location and wait for it to finish completing. The process can take some time to finish depending on the size of the spritesheet. Use the new tile palette to paint the tilemaps.